move graph to graveyard
