+++ packages/grafana-ui/src/components/index.ts
@@ -117,9 +117,6 @@ export {
 export { Sparkline } from './Sparkline/Sparkline';
 
 export { Gauge } from './Gauge/Gauge';
-export { Graph } from './Graph/Graph';
-export { GraphWithLegend } from './Graph/GraphWithLegend';
-export { GraphContextMenu, GraphContextMenuHeader } from './Graph/GraphContextMenu';
 export { BarGauge } from './BarGauge/BarGauge';
 export {
   VizTooltip,
@@ -130,7 +127,6 @@ export {
   type SeriesTableRowProps,
 } from './VizTooltip';
 export { VizRepeater, type VizRepeaterRenderValueProps } from './VizRepeater/VizRepeater';
-export { graphTimeFormat, graphTickFormatter } from './Graph/utils';
 export {
   PanelChrome,
   type PanelChromeProps,
@@ -154,7 +150,7 @@ export { VizLegend } from './VizLegend/VizLegend';
 export { VizLegendListItem } from './VizLegend/VizLegendListItem';
 
 export { Alert, type AlertVariant } from './Alert/Alert';
-export { GraphSeriesToggler, type GraphSeriesTogglerAPI } from './Graph/GraphSeriesToggler';
+export { GraphSeriesToggler, type GraphSeriesTogglerAPI } from '../graveyard/Graph/GraphSeriesToggler';
 export { Collapse, ControlledCollapse } from './Collapse/Collapse';
 export { CollapsableSection } from './Collapse/CollapsableSection';
 export { DataLinkButton } from './DataLinks/DataLinkButton';
@@ -295,6 +291,16 @@ export { type UPlotConfigPrepFn } from './uPlot/config/UPlotConfigBuilder';
 export * from './PanelChrome/types';
 export { Label as BrowserLabel } from './BrowserLabel/Label';
 export { PanelContainer } from './PanelContainer/PanelContainer';
+
+// -----------------------------------------------------
+// Graveyard: exported, but no longer used internally
+// These will be removed in the future
+// -----------------------------------------------------
+
+export { Graph } from '../graveyard/Graph/Graph';
+export { GraphWithLegend } from '../graveyard/Graph/GraphWithLegend';
+export { GraphContextMenu, GraphContextMenuHeader } from '../graveyard/Graph/GraphContextMenu';
+export { graphTimeFormat, graphTickFormatter } from '../graveyard/Graph/utils';
 
 export { GraphNG, type GraphNGProps } from '../graveyard/GraphNG/GraphNG';
 export { TimeSeries } from '../graveyard/TimeSeries/TimeSeries';
