+++ packages/grafana-ui/src/graveyard/Graph/GraphSeriesToggler.tsx
