+++ packages/grafana-ui/src/graveyard/Graph/GraphTooltip/MultiModeGraphTooltip.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import { getValueFromDimension } from '@grafana/data';
 
-import { SeriesTable } from '../../VizTooltip';
-import { FlotPosition } from '../../VizTooltip/VizTooltip';
+import { SeriesTable } from '../../../components/VizTooltip';
+import { FlotPosition } from '../../../components/VizTooltip/VizTooltip';
 import { getMultiSeriesGraphHoverInfo } from '../utils';
 
 import { GraphTooltipContentProps } from './types';
