+++ packages/grafana-ui/src/graveyard/Graph/types.ts
